+++ ceammc/CHANGELOG.md
@@ -38,6 +38,7 @@
 - array.play~ added (with array.p~ alias)
 - env.smooth~ play method added
 - dbfs->amp~ converter added
+- symbol2intlist (with sym->ilist alias) converter added
 - hex->int~ converter added
 - freeze added to fx.freeverb~, fx.freeverb2~ and fx.zita_rev1~ (via @freeze property)
 - flt.ff_comb~ added (feed forward comb filter)